// packages/lego/src/components/DatePickerV2/src/utils/types.ts
export interface NDatePickerDomRefs {
  startDatePanelsDomRef: Ref<
    LegoComponentInstance<typeof DatePanel, DatePickerV2DatePanelExposes> | undefined
  >;
  endDatePanelsDomRef: Ref<
    LegoComponentInstance<typeof DatePanel, DatePickerV2DatePanelExposes> | undefined
  >;
  pickerDomRef: Ref<LegoComponentInstance<typeof NPicker, PickerExposes> | undefined>;
  startInputDomRef: Ref<
    LegoComponentInstance<typeof NPickerPureInput, PickerPureInputExposes> | undefined
  >;
  endInputDomRef: Ref<
    LegoComponentInstance<typeof NPickerPureInput, PickerPureInputExposes> | undefined
  >;
}